Safari-609.1.11.1/LayoutTests/inspector/dom/event-listener-add-remove-expected.txt

Testing events for adding and removing event listeners.

Changing inspected node to #document...

== Running test suite: DOM.eventListeners
-- Running test case: DOM.eventListeners.Initial
PASS: There should be no event listeners.

-- Running test case: DOM.eventListeners.Add
Adding listener "test".
PASS: There should be 1 event listeners.
  - "test"

-- Running test case: DOM.eventListeners.Remove
Removing listener "test".
PASS: There should be 0 event listeners.

-- Running test case: DOM.eventListeners.AddAndRemove
Adding listener "test".
PASS: There should be 1 event listeners.
  - "test"

Removing listener "test".
PASS: There should be 0 event listeners.

-- Running test case: DOM.eventListeners.Once
Adding listener "click".
PASS: There should be 1 event listeners.
  - "click"

Trigger "click" event.
PASS: There should be 0 event listeners.

-- Running test case: DOM.eventListeners.AttributeAddAndRemove
Adding attribute "onload" with callback 1.
PASS: There should be 1 event listeners.
  - "load"

Removing attribute "onload".
PASS: There should be 0 event listeners.

-- Running test case: DOM.eventListeners.AttributeReplace
Adding attribute "onload" with callback 1.
PASS: There should be 1 event listeners.
  - "load"

Adding attribute "onload" with callback 2.
Attribute event listener replaced.
PASS: There should be 1 event listeners.
  - "load"

Removing attribute "onload".
PASS: There should be 0 event listeners.
